Update Home SpreadBooks View
